New 2.8 keymap
This is similar to the old one, but with some key mappings taken from the new Industry Standard keymap. 
Biggest changes are with transformation tools. Now Grab is W, Rotate is E, Scale is R and box select is Q. 
Select/Deselect all is Ctrl + A. Apply menu is Return. 
As well as some other changes.  
I did this to make main fuctions the same as in Industry Standard keymap, so it is easier to use that keymap on computers without this custom keymap. It's now more like extended industry standard keymap. 

Likely from now on, i will use and update only this one.